Use flexbox instead of grid
I get incorrect results in mobile browsers when it comes to using vh (because of the address bar/bottom bar).
I tried various workarounds for a bit and it seems using flexbox is the cleanest solution.

The current version however doesn't work on desktop safari because aspect-ratio isn't yet released.